Fixes crash if continueUserActivity is called without a webpageURL
If your app is opened by`continueUserActivity` in the AppDelegate, but without a webpage url (eg. from a spotlight item / other) the Iterable Kit crashes in the NSRegularExpression call here, which can't take a nil string as a parameter.

Verified that the app continues to launch correctly with this nil check in place.